+++ CHANGELOG.md
@@ -29,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - iBoM: `mark_when_checked` option (#881)
 - KiRi: Support for uninitialized submodules (#888)
 - Report: `buried_vias_count` for KiCad 10
-- SCH Variant: `include` option to export legacy variants to KiCad 10
+- SCH/PCB Variant: `include` option to export legacy variants to KiCad 10
 
 ### Fixed
 - 3D models: download models from the repo tag corresponding to the currently